test: Change erased type in AnnotationTest
Just to test a type erasure, we currently reference
javax.accessibility.AccessibilityProvider, which is in the "java.deskop"
module.

If we were to modularize jsweet, we would have to "requires
java.desktop" just for this (or Eclipse would not be able to run unit
tests).

Moreover, that module may not exist in all JVM environments.

Change to a similar class in "java.base", ServiceLoader, and adjust
module-info.java.txt accordingly.