candy-generator: Build parsers with Java, not Ant
Previously, we were building the used parsers with an Ant script.
This was brittle, especially updating dependency versions was difficult.

Moreover, the generated files contained timestamps, which is why these
files were not included in version control.

Replace the generator builder code with a Java class that gets executed
upon Maven build.

The generator automatically transforms the output to remove all
information about the current timestamp, making it possible to include
the files in version control. This also helps checking for errors when
updating the parser generators.